// src/interfaces/product.interface.ts
export interface Product {
  id: string;
  description: string | null;
  images: string[];
  inStock: number;
  price: number;
  sizes: Size[];
  slug: string;
  tags: string[];
  title: string;
  ProductImage?: ProductImage[];
  //type: ValidTypes;
  gender?: "men" | "women" | "kid" | "unisex" | "male" | "female" | null;
}
export interface CartProduct {
  id: string;
  slug: string;
  price: number;
  quantity: number;
  size: Size;
  image: string;
  title: string;
}

export interface ProductImage {
  url: string;
}

export type Category = "men" | "women" | "kid" | "unisex";
export type Size = "XS" | "S" | "M" | "L" | "XL" | "XXL" | "XXXL";
export type ValidTypes = "shirts" | "pants" | "hoodies" | "hats";